Add Transform module and remove them from Enigma
